move file path gin converter to new file
